feat: return Vec<u8> in BlockingLogIterator
